+++ db/migrate/20150712154833_add_slug_to_news_stories.rb
@@ -1,0 +1,6 @@
+class AddSlugToNewsStories < ActiveRecord::Migration
+  def change
+    add_column :news_stories, :slug, :string
+    add_index :news_stories, :slug, unique: true
+  end
+end