+++ lib/news/engine.rb
@@ -1,5 +1,7 @@
 require 'awesome_nested_set'
 require 'bootstrap-sass'
+require 'friendly_id'
+require 'jquery-rails'
 
 module News
   class Engine < ::Rails::Engine